+++ readium/adapters/pdfium/navigator/src/main/java/org/readium/adapter/pdfium/navigator/PdfiumDefaults.kt
@@ -6,6 +6,7 @@
 
 package org.readium.adapter.pdfium.navigator
 
+import org.readium.r2.navigator.preferences.Axis
 import org.readium.r2.navigator.preferences.ReadingProgression
 
 /**
@@ -19,4 +20,5 @@ public data class PdfiumDefaults(
     val pageSpacing: Double? = null,
     val readingProgression: ReadingProgression? = null,
     val scroll: Boolean? = null,
+    val scrollAxis: Axis? = null,
 )